HeroAI Stability changes
changing caching
