+++ Py4GW.ini
@@ -2,7 +2,7 @@
 py4gw_dll_name = Py4GW.dll
 blackbox_dll_name = GWBlackBOX.dll
 account_config_file = accounts.json
-autoexec_script = C:/Users/Apo/Py4GW_python_files/Py4GW - LDoA.py
+autoexec_script = C:/Users/Apo/Py4GW_python_files/HeroAI.py
 layout_mode = Compact Layout
 
 [Instance Timer]
@@ -27,9 +27,9 @@ subcategory = Explorable
 
 [Widget Manager]
 x = 656
-y = 11
+y = 4
 enable_all = True
-collapsed = False
+collapsed = True
 
 [Enter character on load]
 category = Triggers
