+++ Widgets/Config/Travel.ini
@@ -1,6 +1,6 @@
 [Outpost Travel Config]
-x = 831
-y = 12
+x = 830
+y = 13
 collapsed = True
 config_x = 402
 config_y = 531
