+++ imgui.ini
@@ -5,7 +5,7 @@ Collapsed=0
 
 [Window][Py4GW Console]
 Pos=60,60
-Size=494,436
+Size=250,250
 Collapsed=0
 
 [Window][Shutdown]
@@ -14,8 +14,18 @@ Size=16,35
 Collapsed=0
 
 [Window][Widget Manager]
-Pos=656,11
-Size=212,218
+Pos=656,4
+Size=100,100
+Collapsed=1
+
+[Window][HeroAI - Control Panel]
+Pos=70,498
+Size=265,150
+Collapsed=0
+
+[Window][Multibox Tools]
+Pos=135,329
+Size=150,148
 Collapsed=0
 
 [Window][Intance Timer##Instance Timer]
@@ -24,27 +34,11 @@ Size=155,50
 Collapsed=0
 
 [Window][Outpost Travel]
-Pos=831,12
+Pos=830,13
 Size=300,200
 Collapsed=1
 
-[Window][  TH3KUM1KO'S PRESEARING BIBLE]
-Pos=79,191
-Size=389,1361
-Collapsed=0
-
-[Window][Py4GW Lib DEMO]
-Pos=710,512
-Size=400,360
-Collapsed=0
-
-[Window][PyImGui DEMO]
-Pos=424,497
-Size=460,340
-Collapsed=0
-
-[Window][PyImGui Selectables DEMO]
-Pos=401,440
-Size=300,425
-Collapsed=0
+[Table][0x08B1E730,2]
+RefScale=14
+Column 1  Sort=0v
